+++ packages/Skeleton/stories/Skeleton.stories.tsx
@@ -1,8 +1,7 @@
 import { Meta, Story } from '@storybook/react/types-6-0';
-import React from 'react';
+import React, { useState } from 'react';
 import { Skeleton } from '@blaze-react/skeleton';
 import { preset, ThemeProvider } from '@blaze-react/themes';
-import { controllerArgs, controllerArgTypes } from './shared';
 
 export default {
   title: '@blaze-react/Skeleton',
@@ -30,8 +29,9 @@ export default {
 } as Meta;
 
 const Template: Story = (args) => {
+  const [theme, setTheme] = useState(preset);
   return (
-    <ThemeProvider theme={preset} showSkeleton>
+    <ThemeProvider theme={theme} setTheme={setTheme} showSkeleton>
       <div className="h-48">
         <Skeleton {...args} />
       </div>
@@ -41,6 +41,4 @@ const Template: Story = (args) => {
 
 export const Default = Template.bind({});
 
-Default.args = {
-  theme: preset;
-};
+Default.args = {};