+++ packages/themes/__tests__/themeContext.test.tsx
@@ -1,5 +1,5 @@
-import React, { useState } from 'react';
-import { render, screen } from '@testing-library/react';
+import React from 'react';
+import { render } from '@testing-library/react';
 import { ThemeContext } from '../src/context/themeContext';
 import { ThemeProvider } from '../src/context/themeProvider';
 import { preset } from '@blaze-react/themes';
@@ -8,7 +8,7 @@ import { Button } from '@blaze-react/button';
 describe('Theme Context', () => {
   test('it should render correctly', () => {
     const { asFragment } = render(
-      <ThemeProvider theme={preset}>
+      <ThemeProvider theme={preset} showSkeleton>
         <Button />
       </ThemeProvider>,
     );
@@ -19,7 +19,9 @@ describe('Theme Context', () => {
     const setTheme = jest.fn();
 
     const { asFragment } = render(
-      <ThemeContext.Provider value={{ theme: preset, setTheme: setTheme }}></ThemeContext.Provider>,
+      <ThemeContext.Provider
+        value={{ theme: preset, setTheme: setTheme, showSkeleton: false }}
+      ></ThemeContext.Provider>,
     );
 
     expect(asFragment).toMatchSnapshot();